refactor(orchestrator): LangGraph full design — reco + per-node harness
Reflect graph.md full design. orchestrator.py is now the LangGraph graph
(control flow); heavy execution functions moved to pipeline_steps.py.

- rename ml/orchestrator.py → ml/pipeline_steps.py (shared step library:
  run_cmd, parsers, claude_analyze, stage_*, _fe_train_eval/_build/_commit,
  fe_state io, constants). Strip the old while-loop main/stage_fe/_fe_run.
- new ml/orchestrator.py: LangGraph StateGraph with
  * claude feature recommendation node (n_recommend): weak-scenario diagnosis
    → claude -p proposes/validates new candidate lambdas → dynamic_candidates.py
    → candidate pool; convergence re-recommend loop (invent_rounds cap).
  * per-node claude harness (claude_harness factory, HARNESS_ON toggle):
    analyze each compute node output → verdict continue/retry/stop → routing.
  * fe_baseline node (feature_engineer --diagnose_only) split from fe_train.
  * interrupt() gates for deploy/release/converge (HITL, irreversible-safe).
  * log_sheet(kind) DRY factory for Sheets logging.
- remove ml/orchestrator_lg.py (superseded by this orchestrator.py).
- re-add dynamic_candidates.py consumption in feature_engineer.py so
  recommended features are scanned; gitignore the generated file.

Verified: AST + graph compile (27 nodes) + mock end-to-end run
(2 adoptions + convergence → 3 branches, clean terminate).
